+++ CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - CASM-4350: Implement a proto type of Split config maps, to store cray-product-catalog data...
-- CASM-4227: Implement a prototype to have granular query from main and sub ConfigMap
+- CASM-4427: Implement a prototype to have granular query from main and sub ConfigMap
+- CASM-3981: Add S3 artifacts and loftsman manifests to the Product Catalog Schema
 
 ## [1.8.8] - 2023-05-31